+++ Cargo.toml
@@ -49,7 +49,7 @@ path = "src/lib.rs"
 features = ["cl03"]
 
 [features]
-default = ["bbsplus"]
+default = ["bbsplus", "cl03"]
 cl03 = ["dep:rug"]
 bbsplus = ["dep:bls12_381_plus"]
