+++ src/bbsplus/signature.rs
@@ -16,7 +16,7 @@
 
 use bls12_381_plus::{G1Projective, Scalar, G2Projective, Gt, multi_miller_loop, G2Prepared};
 use serde::{Deserialize, Serialize};
-use crate::{bbsplus::{ciphersuites::BbsCiphersuite, generators::Generators}, errors::Error, schemes::{algorithms::BBSplus, generics::Signature}, utils::{message::BBSplusMessage, util::bbsplus_utils::{calculate_domain, hash_to_scalar, parse_g1_projective, serialize, ScalarExt}}};
+use crate::{bbsplus::{ciphersuites::BbsCiphersuite, generators::Generators}, errors::Error, schemes::{algorithms::BBSplus, generics::Signature}, utils::{message::bbsplus_message::BBSplusMessage, util::bbsplus_utils::{calculate_domain, hash_to_scalar, parse_g1_projective, serialize, ScalarExt}}};
 use elliptic_curve::{hash2curve::ExpandMsg, group::Curve};
 use super::keys::{BBSplusPublicKey, BBSplusSecretKey};
 
@@ -288,4 +288,237 @@ where
     } else {
         Err(Error::SignatureVerificationError)
     }
+}
+
+
+
+
+#[cfg(test)]
+mod tests {
+    
+    use std::fs;
+    use crate::bbsplus::ciphersuites::BbsCiphersuite;
+    use crate::keys::pair::KeyPair;
+    use elliptic_curve::hash2curve::ExpandMsg;
+    use crate::schemes::algorithms::Scheme;
+    use crate::{bbsplus::keys::{BBSplusPublicKey, BBSplusSecretKey}, schemes::{algorithms::BBSplus, generics::Signature}};
+    use crate::schemes::algorithms::{BbsBls12381Sha256, BbsBls12381Shake256};
+    
+
+    //MSG SIGNATURE
+    #[test]
+    fn msg_signature_sha256_1() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature001.json");
+    }
+    #[test]
+    fn msg_signature_sha256_2() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature002.json");
+    }
+    #[test]
+    fn msg_signature_sha256_3() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature003.json");
+    }
+    #[test]
+    fn msg_signature_sha256_4() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature004.json");
+    }
+    #[test]
+    fn msg_signature_sha256_5() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature005.json");
+    }
+    #[test]
+    fn msg_signature_sha256_6() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature006.json");
+    }
+    #[test]
+    fn msg_signature_sha256_7() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature007.json");
+    }
+    #[test]
+    fn msg_signature_sha256_8() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature008.json");
+    }
+    #[test]
+    fn msg_signature_sha256_9() {
+        msg_signature::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/", "signature/signature009.json");
+    }
+
+
+    //MSG SIGNATURE - SHAKE256
+    #[test]
+    fn msg_signature_shake256_1() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature001.json");
+    }
+    #[test]
+    fn msg_signature_shake256_2() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature002.json");
+    }
+    #[test]
+    fn msg_signature_shake256_3() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature003.json");
+    }
+    #[test]
+    fn msg_signature_shake256_4() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature004.json");
+    }
+    #[test]
+    fn msg_signature_shake256_5() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature005.json");
+    }
+    #[test]
+    fn msg_signature_shake256_6() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature006.json");
+    }
+    #[test]
+    fn msg_signature_shake256_7() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature007.json");
+    }
+    #[test]
+    fn msg_signature_shake256_8() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature008.json");
+    }
+    #[test]
+    fn msg_signature_shake256_9() {
+        msg_signature::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/", "signature/signature009.json");
+    }
+
+
+
+    //Update Signature - SHA256
+    #[test]
+    fn update_signature_sha256() {
+        update_signature::<BbsBls12381Sha256>();
+    }
+
+    //Update Blinded Signature - SHAKE256
+    #[test]
+    fn update_signature_shake256() {
+        update_signature::<BbsBls12381Shake256>();
+    }
+
+
+    fn msg_signature<S: Scheme>(pathname: &str, filename: &str) 
+    where
+        S::Ciphersuite: BbsCiphersuite,
+        <S::Ciphersuite as BbsCiphersuite>::Expander: for<'a> ExpandMsg<'a>,
+    {
+        let data = fs::read_to_string([pathname, filename].concat()).expect("Unable to read file");
+        let res: serde_json::Value = serde_json::from_str(&data).expect("Unable to parse");
+        eprintln!("{}", res["caseName"]);
+
+        let header_hex = res["header"].as_str().unwrap();
+        let msgs_hex: Vec<String> = res["messages"].as_array().unwrap().iter().map(|m| serde_json::from_value(m.clone()).unwrap()).collect();
+        let SK_hex = res["signerKeyPair"]["secretKey"].as_str().unwrap();
+        let PK_hex = res["signerKeyPair"]["publicKey"].as_str().unwrap();
+        let SIGNATURE_expected = res["signature"].as_str().unwrap();
+        let RESULT_expected = res["result"]["valid"].as_bool().unwrap();
+
+        let header = hex::decode(header_hex).unwrap();
+        let SK = BBSplusSecretKey::from_bytes(&hex::decode(SK_hex).unwrap()).unwrap();
+        let PK = BBSplusPublicKey::from_bytes(&hex::decode(PK_hex).unwrap()).unwrap();
+
+        let messages: Vec<Vec<u8>> = msgs_hex.iter().map(|m| hex::decode(m).unwrap()).collect();
+
+        let signature = Signature::<BBSplus<S::Ciphersuite>>::sign(Some(&messages), &SK, &PK, Some(&header)).unwrap();
+
+        let result0 = hex::encode(signature.to_bytes()) == SIGNATURE_expected;
+
+        let result1 = result0 == RESULT_expected;
+        if !result1 {
+            eprintln!("  SIGN: {}", result1);
+            eprintln!("  Expected: {}", SIGNATURE_expected);
+            eprintln!("  Computed: {}", hex::encode(signature.to_bytes()));
+            assert!(result1, "failed")
+        }
+
+
+
+        //Verify the signature
+
+        let signature_expected = Signature::<BBSplus<S::Ciphersuite>>::from_bytes(&hex::decode(SIGNATURE_expected).unwrap().try_into().unwrap()).unwrap();
+        let result2 = signature_expected.verify(&PK, Some(&messages),  Some(&header)).is_ok();
+        let result3 = result2 == RESULT_expected;
+
+        if !result3 {
+            eprintln!("  VERIFY: {}", result3);
+            eprintln!("  Expected: {}", RESULT_expected);
+            eprintln!("  Computed: {}", result2);
+            assert!(result3, "failed");
+        
+        }else {
+            eprintln!("  SIGN: {}", result1);
+            eprintln!("  Expected: {}", SIGNATURE_expected);
+            eprintln!("  Computed: {}", hex::encode(signature.to_bytes()));
+        
+            eprintln!("  VERIFY: {}", result3);
+            eprintln!("  Expected: {}", RESULT_expected);
+            eprintln!("  Computed: {}", result2);
+            if RESULT_expected == false {
+                eprintln!("{} ({})", result3, res["result"]["reason"].as_str().unwrap());
+            }
+        }
+
+    }
+
+
+
+    fn update_signature<S: Scheme>()
+    where
+        S::Ciphersuite: BbsCiphersuite,
+        <S::Ciphersuite as BbsCiphersuite>::Expander: for<'a> ExpandMsg<'a>,
+    {
+        const IKM: &str = "746869732d49532d6a7573742d616e2d546573742d494b4d2d746f2d67656e65726174652d246528724074232d6b6579";
+        const KEY_INFO: &str = "746869732d49532d736f6d652d6b65792d6d657461646174612d746f2d62652d757365642d696e2d746573742d6b65792d67656e";
+        const msgs: [&str; 3] = ["9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02", "87a8bd656d49ee07b8110e1d8fd4f1dcef6fb9bc368c492d9bc8c4f98a739ac6", "96012096adda3f13dd4adbe4eea481a4c4b5717932b73b00e31807d3c5894b90"];
+        const header_hex: &str = "11223344556677889900aabbccddeeff";
+        let header = hex::decode(header_hex).unwrap();
+
+        let keypair = KeyPair::<BBSplus<S::Ciphersuite>>::generate(
+            &hex::decode(&IKM).unwrap(),
+            Some(&hex::decode(&KEY_INFO).unwrap()),
+            None
+        ).unwrap();
+
+        let sk = keypair.private_key();
+        let pk = keypair.public_key();
+
+        let messages: Vec<Vec<u8>> = msgs.iter().map(|m| hex::decode(m).unwrap()).collect();
+
+        let signature = Signature::<BBSplus<S::Ciphersuite>>::sign(Some(&messages), sk, pk, Some(&header)).unwrap();
+        let verify = signature.verify(pk, Some(&messages), Some(&header)).is_ok();
+
+        assert!(verify, "Signature NOT VALID!");
+
+
+        const new_message: &str = "8872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02";
+        const update_index: usize = 0usize;
+
+        let new_message_bytes = hex::decode(new_message).unwrap();
+        let old_message_bytes = messages.get(update_index).unwrap();
+
+
+        // let new_message_scalar = BBSplusMessage::map_message_to_scalar_as_hash::<S::Ciphersuite>(&hex::decode(new_message).unwrap(), Some(&dst)).unwrap();
+        // let old_message_scalar = msgs_scalars.get(update_index).unwrap();
+
+        let updated_signature = signature.update_signature(sk, &old_message_bytes, &new_message_bytes, update_index, messages.len()).unwrap();
+
+        let mut new_messages = messages.clone();
+        new_messages[update_index] = new_message_bytes;
+
+        let verify = updated_signature.verify(pk, Some(&new_messages), Some(&header)).is_ok();
+
+        assert!(verify, "Signature NOT VALID!");
+
+
+        const new_message_wrong: &str = "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f01";
+        let new_message_bytes_wrong= hex::decode(new_message_wrong).unwrap();
+        let mut new_messages_wrong = messages.clone();
+        new_messages_wrong[update_index] = new_message_bytes_wrong;
+
+        let verify = updated_signature.verify(pk, Some(&new_messages_wrong), Some(&header)).is_ok();
+
+        assert!(!verify, "Signature MUST BE NOT VALID!");
+
+    }
+
 }
