+++ src/cl03/blind.rs
@@ -17,7 +17,7 @@ use std::panic;
 use digest::Digest;
 use rug::{Integer, ops::Pow};
 use serde::{Deserialize, Serialize};
-use crate::{schemes::algorithms::CL03, utils::message::CL03Message, cl03::{ciphersuites::CLCiphersuite, bases::Bases}, utils::random::{random_prime, random_bits}, schemes::generics::{BlindSignature, Commitment, Signature, ZKPoK}};
+use crate::{schemes::algorithms::CL03, utils::message::cl03_message::CL03Message, cl03::{ciphersuites::CLCiphersuite, bases::Bases}, utils::random::{random_prime, random_bits}, schemes::generics::{BlindSignature, Commitment, Signature, ZKPoK}};
 use super::{keys::{CL03SecretKey, CL03PublicKey, CL03CommitmentPublicKey}, commitment::CL03Commitment, signature::CL03Signature};
 
 
@@ -106,4 +106,118 @@ impl <CS:CLCiphersuite> BlindSignature<CL03<CS>> {
         let sig = CL03BlindSignature{e: self.e().clone(), rprime: self.rprime().clone(), v};
         Self::CL03(sig)
     }
+}
+
+
+
+#[cfg(test)]
+mod tests {
+
+    use digest::Digest;
+    use crate::{keys::pair::KeyPair, schemes::algorithms::{Scheme, CL03, Ciphersuite}, utils::message::cl03_message::CL03Message, schemes::generics::{Commitment, ZKPoK, BlindSignature}, cl03::bases::Bases};
+    use crate::cl03::ciphersuites::CLCiphersuite;
+    use crate::schemes::algorithms::CL03_CL1024_SHA256;
+
+
+
+    //Blind signature - CL1024-SHA256
+    #[test]
+    fn blind_sign_cl1024_sha256() {
+        blind_sign::<CL03_CL1024_SHA256>();
+    }
+
+
+    //Blind Signature update - CL1024-SHA256
+    #[test]
+    fn update_signature_cl1024_sha256() {
+        update_signature::<CL03_CL1024_SHA256>();
+    }
+
+
+    fn blind_sign<S: Scheme>() 
+    where
+        S::Ciphersuite: CLCiphersuite,
+        <S::Ciphersuite as Ciphersuite>::HashAlg: Digest
+    {
+
+        const msgs: &[&str] = &["9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02", "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f03", "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f04"];
+        // const msg: &str = "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02";
+        const wrong_msgs: &[&str] = &["7872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02", "7872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f03", "7872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f04"];
+
+        let cl03_keypair = KeyPair::<CL03<S::Ciphersuite>>::generate();
+        let a_bases = Bases::generate(cl03_keypair.public_key(), msgs.len());
+
+        let messages: Vec<CL03Message> = msgs.iter().map(|&m| CL03Message::map_message_to_integer_as_hash::<S::Ciphersuite>(&hex::decode(m).unwrap()) ).collect();
+        // let msg_intger = CL03Message::map_message_to_integer_as_hash::<S::Ciphersuite>(&hex::decode(msg).unwrap());
+        // let messages = [msg_intger.clone()];
+        let wrong_messages: Vec<CL03Message> = wrong_msgs.iter().map(|&m| CL03Message::map_message_to_integer_as_hash::<S::Ciphersuite>(&hex::decode(m).unwrap()) ).collect();
+
+    
+        let unrevealed_message_indexes = [0usize];
+        let revealed_message_indexes = [1usize,  2usize];
+        let revealed_messages: Vec<CL03Message> = messages.iter().enumerate().filter(|&(i,_)| revealed_message_indexes.contains(&i) ).map(|(_, m)| m.clone()).collect();
+        let revealed_messages_wrong : Vec<CL03Message> = wrong_messages.iter().enumerate().filter(|&(i,_)| revealed_message_indexes.contains(&i) ).map(|(_, m)| m.clone()).collect();
+
+        let commitment = Commitment::<CL03<S::Ciphersuite>>::commit_with_pk(&messages, cl03_keypair.public_key(), &a_bases, Some(&unrevealed_message_indexes));
+
+        let zkpok = ZKPoK::<CL03<S::Ciphersuite>>::generate_proof(&messages, commitment.cl03Commitment(), None, cl03_keypair.public_key(), &a_bases, None, &unrevealed_message_indexes);
+
+        let blind_signature = BlindSignature::<CL03<S::Ciphersuite>>::blind_sign(cl03_keypair.public_key(), cl03_keypair.private_key(), &a_bases, &zkpok, Some(&revealed_messages), commitment.cl03Commitment(), None, None, &unrevealed_message_indexes, Some(&revealed_message_indexes));
+        let unblided_signature = blind_signature.unblind_sign(&commitment);
+        let verify = unblided_signature.verify_multiattr(cl03_keypair.public_key(), &a_bases, &messages);
+
+        assert!(verify, "Error! The unblided signature verification should PASS!");
+
+
+        let blind_signature_wrong = BlindSignature::<CL03<S::Ciphersuite>>::blind_sign(cl03_keypair.public_key(), cl03_keypair.private_key(), &a_bases, &zkpok, Some(&revealed_messages_wrong), commitment.cl03Commitment(), None, None, &unrevealed_message_indexes, Some(&revealed_message_indexes));
+        let unblided_signature_wrong = blind_signature_wrong.unblind_sign(&commitment);
+        let verify = unblided_signature_wrong.verify_multiattr(cl03_keypair.public_key(), &a_bases, &messages);
+
+        assert!(!verify, "Error! The unblinded signature verification SHOULD FAIL!");
+
+
+    }
+
+
+    pub(crate) fn update_signature<S: Scheme>()
+    where
+        S::Ciphersuite: CLCiphersuite,
+        <S::Ciphersuite as Ciphersuite>::HashAlg: Digest
+    {
+        const msgs: &[&str] = &["9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02", "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f03", "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f04"];
+        const updated_msgs: &[&str] = &["9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f02", "7872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f03", "9872ad089e452c7b6e283dfac2a80d58e8d0ff71cc4d5e310a1debdda4a45f04"];
+
+        let n_attr = msgs.len();
+        let cl03_keypair = KeyPair::<CL03<S::Ciphersuite>>::generate();
+        let a_bases = Bases::generate(cl03_keypair.public_key(), n_attr);
+
+
+        let messages: Vec<CL03Message> = msgs.iter().map(|&m| CL03Message::map_message_to_integer_as_hash::<S::Ciphersuite>(&hex::decode(m).unwrap()) ).collect();
+        let updated_messages: Vec<CL03Message> = updated_msgs.iter().map(|&m| CL03Message::map_message_to_integer_as_hash::<S::Ciphersuite>(&hex::decode(m).unwrap()) ).collect();
+    
+        let unrevealed_message_indexes = [0usize];
+        let revealed_message_indexes = [1usize,  2usize];
+        let revealed_messages: Vec<CL03Message> = messages.iter().enumerate().filter(|&(i,_)| revealed_message_indexes.contains(&i) ).map(|(_, m)| m.clone()).collect();
+        let revealed_updated_messages: Vec<CL03Message> = updated_messages.iter().enumerate().filter(|&(i,_)| revealed_message_indexes.contains(&i) ).map(|(_, m)| m.clone()).collect();
+        let commitment = Commitment::<CL03<S::Ciphersuite>>::commit_with_pk(&messages, cl03_keypair.public_key(), &a_bases, Some(&unrevealed_message_indexes));
+        
+        let zkpok = ZKPoK::<CL03<S::Ciphersuite>>::generate_proof(&messages, commitment.cl03Commitment(), None, cl03_keypair.public_key(), &a_bases, None, &unrevealed_message_indexes);
+
+        let blind_signature = BlindSignature::<CL03<S::Ciphersuite>>::blind_sign(cl03_keypair.public_key(), cl03_keypair.private_key(), &a_bases, &zkpok, Some(&revealed_messages), commitment.cl03Commitment(), None, None, &unrevealed_message_indexes, Some(&revealed_message_indexes));
+        let unblided_signature = blind_signature.unblind_sign(&commitment);
+        let verify = unblided_signature.verify_multiattr(cl03_keypair.public_key(), &a_bases, &messages);
+
+        assert!(verify, "Error! The unblided signature verification should PASS!");
+
+
+
+        let updated_signature = blind_signature.update_signature(Some(&revealed_updated_messages), &commitment.cl03Commitment(), cl03_keypair.private_key(), cl03_keypair.public_key(), &a_bases, Some(&revealed_message_indexes));
+        let unblinded_updated_signature = updated_signature.unblind_sign(&commitment);
+
+        let verify = unblinded_updated_signature.verify_multiattr(cl03_keypair.public_key(), &a_bases, &updated_messages);
+        assert!(verify, "Error! The unblided signature verification should PASS!");
+
+
+    }
+
 }
