+++ src/cl03/sigma_protocols.rs
@@ -15,7 +15,7 @@
 use digest::Digest;
 use rug::{Integer, integer::Order, Complete};
 use serde::{Deserialize, Serialize};
-use crate::{utils::{message::CL03Message, random::random_bits, util::cl03_utils::divm}, schemes::algorithms::CL03, schemes::generics::Commitment};
+use crate::{utils::{message::cl03_message::CL03Message, random::random_bits, util::cl03_utils::divm}, schemes::algorithms::CL03, schemes::generics::Commitment};
 use super::{ciphersuites::CLCiphersuite, bases::Bases, keys::{CL03PublicKey, CL03CommitmentPublicKey}, commitment::CL03Commitment, signature::CL03Signature};
