+++ src/utils/message.rs
@@ -13,132 +13,234 @@
 // limitations under the License.
 
 
-use elliptic_curve::hash2curve::ExpandMsg;
-use serde::{Serialize, Deserialize};
-use crate::errors::Error;
-
-
-#[cfg(feature = "cl03")]
-use crate::cl03::ciphersuites::CLCiphersuite;
-#[cfg(feature = "cl03")]
-use digest::Digest;
-#[cfg(feature = "cl03")]
-use rug::{Integer, integer::Order};
-
 
 
 #[cfg(feature = "bbsplus")]
-use bls12_381_plus::Scalar;
-#[cfg(feature = "bbsplus")]
-use crate::bbsplus::ciphersuites::BbsCiphersuite;
-#[cfg(feature = "bbsplus")]
-use super::util::bbsplus_utils::hash_to_scalar;
-
-
-#[cfg(feature = "bbsplus")]
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub struct BBSplusMessage{
-    pub value: Scalar
-}
-
-#[cfg(feature = "bbsplus")]
-impl BBSplusMessage {
-
-    pub fn new(msg: Scalar) -> Self{
-        Self{value: msg}
-    }
-
-
-    /// https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-bbs-signatures-05#name-messages-to-scalars
-    /// 
-    /// # Description
-    /// The messages_to_scalars operation is used to map a list of messages to their respective scalar values
-    /// 
-    /// # Inputs:
-    /// * `messages` (REQUIRED), a vector of octet strings.
-    /// * `api_id` (REQUIRED), octet string. It could be an empty octet string
-    /// 
-    /// # Output:
-    /// * a vector of [`BBSplusMessage`], which is a wrapper to a `Scalar` or [`Error`].
-    /// 
-    pub fn messages_to_scalar<CS: BbsCiphersuite>(messages: &[Vec<u8>], api_id: &[u8]) -> Result<Vec<Self>, Error>
-    where
-        CS::Expander: for<'a> ExpandMsg<'a>
-    {
-
-        let map_dst = [api_id, CS::MAP_MSG_SCALAR].concat();
-        let mut msg_scalars: Vec<Self> = Vec::new();
-        for m in messages {
-            let scalar = hash_to_scalar::<CS>(m, &map_dst)?;
-            msg_scalars.push(Self { value: scalar })
-        }
-
-        Ok(msg_scalars)
-    }
-
-
+pub mod bbsplus_message {
  
-    /// # Description
-    /// The `map_message_to_scalar_as_hash` operation is used to map a single message to its respective scalar value
-    /// 
-    /// # Inputs:
-    /// * `data` (REQUIRED), an octet string representing a single message.
-    /// * `api_id` (REQUIRED), octet string. It could be an empty octet string
-    /// 
-    /// # Output:
-    /// * a [`BBSplusMessage`], which is a wrapper to a `Scalar` or [`Error`].
-    ///
-    pub fn map_message_to_scalar_as_hash<CS: BbsCiphersuite>(data: &[u8], api_id: &[u8]) -> Result<Self, Error> 
-    where
-        CS::Expander: for<'a> ExpandMsg<'a>,
-    {
+    use bls12_381_plus::Scalar;
+    use elliptic_curve::hash2curve::ExpandMsg;
+    use serde::{Deserialize, Serialize};
+    use crate::bbsplus::ciphersuites::BbsCiphersuite;
+    use crate::errors::Error;
+    use crate::utils::util::bbsplus_utils::hash_to_scalar;
 
-        let map_dst = [api_id, CS::MAP_MSG_SCALAR].concat();
-        let scalar = hash_to_scalar::<CS>(data, &map_dst)?;
-        
-        Ok(Self { value: scalar })
-
+    #[derive(Clone, Copy, PartialEq, Eq, Debug, Serialize, Deserialize)]
+    pub struct BBSplusMessage{
+        pub value: Scalar
     }
 
-    pub fn to_bytes_be(&self) -> [u8; Scalar::BYTES] {
-        self.value.to_be_bytes()
-    }
+    impl BBSplusMessage {
 
-    pub fn from_bytes_be(bytes: &[u8; Scalar::BYTES]) -> Result<Self, Error> {
-        let s = Scalar::from_be_bytes(bytes);
-        if s.is_none().into() {
-            return Err(Error::Unspecified)
+        pub fn new(msg: Scalar) -> Self{
+            Self{value: msg}
         }
-        Ok(BBSplusMessage{value: s.unwrap()})
+
+
+        /// https://datatracker.ietf.org/doc/html/draft-irtf-cfrg-bbs-signatures-05#name-messages-to-scalars
+        /// 
+        /// # Description
+        /// The messages_to_scalars operation is used to map a list of messages to their respective scalar values
+        /// 
+        /// # Inputs:
+        /// * `messages` (REQUIRED), a vector of octet strings.
+        /// * `api_id` (REQUIRED), octet string. It could be an empty octet string
+        /// 
+        /// # Output:
+        /// * a vector of [`BBSplusMessage`], which is a wrapper to a `Scalar` or [`Error`].
+        /// 
+        pub fn messages_to_scalar<CS: BbsCiphersuite>(messages: &[Vec<u8>], api_id: &[u8]) -> Result<Vec<Self>, Error>
+        where
+            CS::Expander: for<'a> ExpandMsg<'a>
+        {
+
+            let map_dst = [api_id, CS::MAP_MSG_SCALAR].concat();
+            let mut msg_scalars: Vec<Self> = Vec::new();
+            for m in messages {
+                let scalar = hash_to_scalar::<CS>(m, &map_dst)?;
+                msg_scalars.push(Self { value: scalar })
+            }
+
+            Ok(msg_scalars)
+        }
+
+
+    
+        /// # Description
+        /// The `map_message_to_scalar_as_hash` operation is used to map a single message to its respective scalar value
+        /// 
+        /// # Inputs:
+        /// * `data` (REQUIRED), an octet string representing a single message.
+        /// * `api_id` (REQUIRED), octet string. It could be an empty octet string
+        /// 
+        /// # Output:
+        /// * a [`BBSplusMessage`], which is a wrapper to a `Scalar` or [`Error`].
+        ///
+        pub fn map_message_to_scalar_as_hash<CS: BbsCiphersuite>(data: &[u8], api_id: &[u8]) -> Result<Self, Error> 
+        where
+            CS::Expander: for<'a> ExpandMsg<'a>,
+        {
+
+            let map_dst = [api_id, CS::MAP_MSG_SCALAR].concat();
+            let scalar = hash_to_scalar::<CS>(data, &map_dst)?;
+            
+            Ok(Self { value: scalar })
+
+        }
+
+        pub fn to_bytes_be(&self) -> [u8; Scalar::BYTES] {
+            self.value.to_be_bytes()
+        }
+
+        pub fn from_bytes_be(bytes: &[u8; Scalar::BYTES]) -> Result<Self, Error> {
+            let s = Scalar::from_be_bytes(bytes);
+            if s.is_none().into() {
+                return Err(Error::Unspecified)
+            }
+            Ok(BBSplusMessage{value: s.unwrap()})
+        }
+
     }
 
+
+
+
+    #[cfg(test)]
+    mod tests {
+        
+        use std::fs;
+        use crate::bbsplus::ciphersuites::BbsCiphersuite;
+        use elliptic_curve::hash2curve::ExpandMsg;
+        use crate::schemes::algorithms::Scheme;
+        use crate::utils::message::bbsplus_message::BBSplusMessage;
+        use crate::schemes::algorithms::{BbsBls12381Sha256, BbsBls12381Shake256};
+        
+
+
+
+        //MAP MESSAGE TO SCALAR - SHA256
+
+        #[test]
+        fn map_message_to_scalar_as_hash_sha256() {
+            map_message_to_scalar_as_hash::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/MapMessageToScalarAsHash.json");
+        }
+
+        //MESSAGES TO SCALAR - SHAKE256
+
+        #[test]
+        fn map_message_to_scalar_as_hash_shake256() {
+            map_message_to_scalar_as_hash::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/MapMessageToScalarAsHash.json");
+        }
+
+
+        #[test]
+        fn messages_to_scalars_sha256() {
+            messages_to_scalars::<BbsBls12381Sha256>("./fixture_data/bls12-381-sha-256/MapMessageToScalarAsHash.json");
+        }
+
+        //MESSAGES TO SCALAR - SHAKE256
+
+        #[test]
+        fn messages_to_scalars_shake256() {
+            messages_to_scalars::<BbsBls12381Shake256>("./fixture_data/bls12-381-shake-256/MapMessageToScalarAsHash.json");
+        }
+
+
+        fn map_message_to_scalar_as_hash<S: Scheme>(filename: &str) 
+        where
+            S::Ciphersuite: BbsCiphersuite,
+            <S::Ciphersuite as BbsCiphersuite>::Expander: for<'a> ExpandMsg<'a>,
+        {
+            let data = fs::read_to_string(filename).expect("Unable to read file");
+            let json: serde_json::Value = serde_json::from_str(&data).expect("Unable to parse");
+            eprintln!("{}", json["caseName"]);
+            let cases = json["cases"].as_array().unwrap();
+
+            let mut boolean = true;
+            for c in cases {
+                let msg = &c["message"];
+
+                let msg_hex = hex::decode(msg.as_str().unwrap()).unwrap();
+
+                let out = hex::encode(BBSplusMessage::map_message_to_scalar_as_hash::<S::Ciphersuite>(&msg_hex, <S::Ciphersuite as BbsCiphersuite>::API_ID).unwrap().to_bytes_be());
+                let out_expected = c["scalar"].as_str().unwrap();
+
+                if out != out_expected{
+                    boolean = false;
+                };
+            }
+
+            assert_eq!(boolean, true);
+        }
+
+
+        fn messages_to_scalars<S: Scheme>(filename: &str) 
+        where
+            S::Ciphersuite: BbsCiphersuite,
+            <S::Ciphersuite as BbsCiphersuite>::Expander: for<'a> ExpandMsg<'a>,
+        {
+            let data = fs::read_to_string(filename).expect("Unable to read file");
+            let json: serde_json::Value = serde_json::from_str(&data).expect("Unable to parse");
+            eprintln!("{}", json["caseName"]);
+            let mut messages = Vec::new();
+            let mut scalars = Vec::new();
+            json["cases"].as_array().unwrap().iter().for_each(|v| {
+                let msg = v["message"].as_str().unwrap();
+                messages.push(hex::decode(msg).unwrap());
+                let s = v["scalar"].as_str().unwrap();
+                scalars.push(s.to_string());
+            });
+
+
+            let message_scalars = BBSplusMessage::messages_to_scalar::<S::Ciphersuite>(&messages, <S::Ciphersuite as BbsCiphersuite>::API_ID).unwrap();
+            
+            let message_scalars_hex: Vec<String> = message_scalars.iter().map(|s| hex::encode(s.to_bytes_be())).collect();
+
+            assert_eq!(scalars, message_scalars_hex);
+
+        }
+
+    }
 }
+
 
 #[cfg(feature = "cl03")]
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub struct CL03Message{
-    pub value: Integer
+pub mod cl03_message {
+
+    use serde::{Serialize, Deserialize};
+    use crate::cl03::ciphersuites::CLCiphersuite;
+    use digest::Digest;
+    use rug::{Integer, integer::Order};
+
+    
+    #[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
+    pub struct CL03Message{
+        pub value: Integer
+    }
+
+
+    impl CL03Message {
+
+        pub fn new(msg: Integer) -> Self {
+            Self{value: msg}
+        }
+
+        pub fn get_value(&self) -> Integer {
+            self.value.clone()
+        }
+
+        pub fn map_message_to_integer_as_hash<C: CLCiphersuite>(data: &[u8]) -> Self 
+        where C::HashAlg: Digest
+        {
+            let binding = <C::HashAlg as Digest>::digest(data);
+            let msg_digest = binding.as_slice();
+            let msg_integer = Integer::from_digits(msg_digest, Order::MsfBe);
+            Self{value: msg_integer}
+
+        }
+
+    }
+
+
+
 }
-#[cfg(feature = "cl03")]
-impl CL03Message {
-
-    pub fn new(msg: Integer) -> Self {
-        Self{value: msg}
-    }
-
-    pub fn get_value(&self) -> Integer {
-        self.value.clone()
-    }
-
-    pub fn map_message_to_integer_as_hash<C: CLCiphersuite>(data: &[u8]) -> Self 
-    where C::HashAlg: Digest
-    {
-        let binding = <C::HashAlg as Digest>::digest(data);
-        let msg_digest = binding.as_slice();
-        let msg_integer = Integer::from_digits(msg_digest, Order::MsfBe);
-        Self{value: msg_integer}
-
-    }
-
-}
-
